+++ Source/Applications/Desktop/Vodovoz/CreateProjectParam.cs
@@ -1,4 +1,4 @@
-using Autofac;
+﻿using Autofac;
 using QS.Dialog.Gtk;
 using QS.Permissions;
 using QS.Project.DB;
@@ -182,6 +182,8 @@ using UserView = Vodovoz.Views.Users.UserView;
 using Vodovoz.ViewModels.ReportsParameters.Payments;
 using Vodovoz.ReportsParameters.Payments;
 using Vodovoz.ViewModels.ViewModels.Reports.Logistics;
+using Vodovoz.ViewModels.ReportsParameters.Production;
+using Vodovoz.Views.ReportsParameters.Production;
 
 namespace Vodovoz
 {
@@ -481,6 +483,7 @@ namespace Vodovoz
 				.RegisterWidgetForWidgetViewModel<NomenclatureOnlineCatalogViewModel, NomenclatureOnlineCatalogView>()
 				.RegisterWidgetForWidgetViewModel<DriversWarehousesEventsJournalFilterViewModel, DriversWarehousesEventsJournalFilterView>()
 				.RegisterWidgetForWidgetViewModel<CompletedDriversWarehousesEventsJournalFilterViewModel, CompletedDriversWarehousesEventsJournalFilterView>()
+				.RegisterWidgetForWidgetViewModel<ProducedProductionReportViewModel, ProducedProductionReportView>()
 				;
 
 			DialogHelper.FilterWidgetResolver = ViewModelWidgetResolver.Instance;
