--- Source/Applications/Desktop/Vodovoz/ReportsParameters/Production/ProducedProductionReport.ProducedProductionReportMode.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Vodovoz.ReportsParameters.Production
-{
-	public partial class ProducedProductionReport
-	{
-		public enum ProducedProductionReportMode
-		{
-			[Display(Name ="Месяц")]
-			Month,
-			[Display(Name = "Год")]
-			Year
-		}
-	}
-}
