--- Source/Applications/Desktop/Vodovoz/gtk-gui/Vodovoz.ReportsParameters.Production.ProducedProductionReport.cs
@@ -1,140 +1,0 @@
-﻿
-// This file has been generated by the GUI designer. Do not modify.
-namespace Vodovoz.ReportsParameters.Production
-{
-	public partial class ProducedProductionReport
-	{
-		private global::Gtk.VBox vbox1;
-
-		private global::Gtk.HBox hbox5;
-
-		private global::Gamma.GtkWidgets.yLabel ylabel3;
-
-		private global::Gamma.Widgets.yEnumComboBox yenumcomboboxMonths;
-
-		private global::Gamma.GtkWidgets.yLabel ylabel5;
-
-		private global::Gamma.Widgets.ySpecComboBox ylistcomboboxYear;
-
-		private global::Gamma.GtkWidgets.yLabel ylabelReportType;
-
-		private global::Gamma.Widgets.yEnumComboBox yenumcomboboxReportType;
-
-		private global::Gtk.VBox vboxParameters;
-
-		private global::Gtk.Button buttonCreateReport;
-
-		protected virtual void Build()
-		{
-			global::Stetic.Gui.Initialize(this);
-			// Widget Vodovoz.ReportsParameters.Production.ProducedProductionReport
-			global::Stetic.BinContainer.Attach(this);
-			this.Name = "Vodovoz.ReportsParameters.Production.ProducedProductionReport";
-			// Container child Vodovoz.ReportsParameters.Production.ProducedProductionReport.Gtk.Container+ContainerChild
-			this.vbox1 = new global::Gtk.VBox();
-			this.vbox1.Name = "vbox1";
-			this.vbox1.Spacing = 6;
-			// Container child vbox1.Gtk.Box+BoxChild
-			this.hbox5 = new global::Gtk.HBox();
-			this.hbox5.Name = "hbox5";
-			this.hbox5.Spacing = 6;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.ylabel3 = new global::Gamma.GtkWidgets.yLabel();
-			this.ylabel3.Name = "ylabel3";
-			this.ylabel3.LabelProp = global::Mono.Unix.Catalog.GetString("Месяц:");
-			this.hbox5.Add(this.ylabel3);
-			global::Gtk.Box.BoxChild w1 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.ylabel3]));
-			w1.Position = 0;
-			w1.Expand = false;
-			w1.Fill = false;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.yenumcomboboxMonths = new global::Gamma.Widgets.yEnumComboBox();
-			this.yenumcomboboxMonths.Name = "yenumcomboboxMonths";
-			this.yenumcomboboxMonths.ShowSpecialStateAll = false;
-			this.yenumcomboboxMonths.ShowSpecialStateNot = false;
-			this.yenumcomboboxMonths.UseShortTitle = false;
-			this.yenumcomboboxMonths.DefaultFirst = false;
-			this.hbox5.Add(this.yenumcomboboxMonths);
-			global::Gtk.Box.BoxChild w2 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.yenumcomboboxMonths]));
-			w2.Position = 1;
-			w2.Expand = false;
-			w2.Fill = false;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.ylabel5 = new global::Gamma.GtkWidgets.yLabel();
-			this.ylabel5.Name = "ylabel5";
-			this.ylabel5.Xalign = 1F;
-			this.ylabel5.LabelProp = "Год:";
-			this.ylabel5.WidthChars = 0;
-			this.ylabel5.MaxWidthChars = 0;
-			this.hbox5.Add(this.ylabel5);
-			global::Gtk.Box.BoxChild w3 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.ylabel5]));
-			w3.Position = 3;
-			w3.Expand = false;
-			w3.Fill = false;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.ylistcomboboxYear = new global::Gamma.Widgets.ySpecComboBox();
-			this.ylistcomboboxYear.Name = "ylistcomboboxYear";
-			this.ylistcomboboxYear.AddIfNotExist = false;
-			this.ylistcomboboxYear.DefaultFirst = false;
-			this.ylistcomboboxYear.ShowSpecialStateAll = false;
-			this.ylistcomboboxYear.ShowSpecialStateNot = false;
-			this.hbox5.Add(this.ylistcomboboxYear);
-			global::Gtk.Box.BoxChild w4 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.ylistcomboboxYear]));
-			w4.Position = 4;
-			w4.Expand = false;
-			w4.Fill = false;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.ylabelReportType = new global::Gamma.GtkWidgets.yLabel();
-			this.ylabelReportType.Name = "ylabelReportType";
-			this.ylabelReportType.LabelProp = global::Mono.Unix.Catalog.GetString("Тип отчета:");
-			this.hbox5.Add(this.ylabelReportType);
-			global::Gtk.Box.BoxChild w5 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.ylabelReportType]));
-			w5.Position = 5;
-			w5.Expand = false;
-			w5.Fill = false;
-			// Container child hbox5.Gtk.Box+BoxChild
-			this.yenumcomboboxReportType = new global::Gamma.Widgets.yEnumComboBox();
-			this.yenumcomboboxReportType.Name = "yenumcomboboxReportType";
-			this.yenumcomboboxReportType.ShowSpecialStateAll = false;
-			this.yenumcomboboxReportType.ShowSpecialStateNot = false;
-			this.yenumcomboboxReportType.UseShortTitle = false;
-			this.yenumcomboboxReportType.DefaultFirst = false;
-			this.hbox5.Add(this.yenumcomboboxReportType);
-			global::Gtk.Box.BoxChild w6 = ((global::Gtk.Box.BoxChild)(this.hbox5[this.yenumcomboboxReportType]));
-			w6.Position = 6;
-			w6.Expand = false;
-			w6.Fill = false;
-			this.vbox1.Add(this.hbox5);
-			global::Gtk.Box.BoxChild w7 = ((global::Gtk.Box.BoxChild)(this.vbox1[this.hbox5]));
-			w7.Position = 0;
-			w7.Expand = false;
-			w7.Fill = false;
-			// Container child vbox1.Gtk.Box+BoxChild
-			this.vboxParameters = new global::Gtk.VBox();
-			this.vboxParameters.Name = "vboxParameters";
-			this.vboxParameters.Spacing = 6;
-			this.vbox1.Add(this.vboxParameters);
-			global::Gtk.Box.BoxChild w8 = ((global::Gtk.Box.BoxChild)(this.vbox1[this.vboxParameters]));
-			w8.Position = 1;
-			// Container child vbox1.Gtk.Box+BoxChild
-			this.buttonCreateReport = new global::Gtk.Button();
-			this.buttonCreateReport.Sensitive = false;
-			this.buttonCreateReport.CanFocus = true;
-			this.buttonCreateReport.Name = "buttonCreateReport";
-			this.buttonCreateReport.UseUnderline = true;
-			this.buttonCreateReport.Label = global::Mono.Unix.Catalog.GetString("Сформировать отчет");
-			this.vbox1.Add(this.buttonCreateReport);
-			global::Gtk.Box.BoxChild w9 = ((global::Gtk.Box.BoxChild)(this.vbox1[this.buttonCreateReport]));
-			w9.PackType = ((global::Gtk.PackType)(1));
-			w9.Position = 2;
-			w9.Expand = false;
-			w9.Fill = false;
-			this.Add(this.vbox1);
-			if ((this.Child != null))
-			{
-				this.Child.ShowAll();
-			}
-			this.Hide();
-		}
-	}
-}
