+++ Source/Applications/Desktop/VodovozViewModels/ReportsParameters/Production/ProducedProductionReportViewModel.ProducedProductionReportMode.cs
@@ -1,0 +1,16 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Vodovoz.ViewModels.ReportsParameters.Production
+{
+	public partial class ProducedProductionReportViewModel
+	{
+		public enum ProducedProductionReportMode
+		{
+			[Display(Name = "Месяц")]
+			Month,
+			[Display(Name = "Год")]
+			Year
+		}
+
+	}
+}
